fix: retry the type detection without the configured type restriction
`getIoBrokerDeviceStates` retried detection without `allowedTypes` when the
restricted attempt found nothing, but handed the detector the options object the
first attempt had already used. `ChannelDetector.detect` caches its pattern list
on that object and records every pattern it checked there, so the retry received
a list frozen under the first attempt's `allowedTypes` with every entry already
marked as checked, and could only ever report nothing. Each attempt now builds
its own options. A device whose states do not describe the type it is configured
as is detected as what those states do describe, which surfaces as the type
mismatch it is instead of passing as a silent debug line.

The three detection entry points move from `MatterAdapter` into
`src/lib/deviceDetection.ts` as free functions over a `DetectionObjectProvider`,
so the detection path can run against an object tree without an adapter
instance. `MatterAdapter` satisfies the interface structurally.

Also fixed, all found by putting that path under test:

- A single-state RGB or RGBW device threw on a colour it could not parse, and
  the throw propagated out of the converter's `init()`, so the bridge dropped the
  device. An unreadable colour now reports black and is logged once per value,
  the way `Cie` already handles one.
- A bridged device whose `init()` failed kept the endpoints it had already added
  to the aggregator, where they answered controllers with cluster defaults and no
  ioBroker state behind them. They are now removed, as they already were when
  adding them failed.
- `DEVICE_DEFAULT_NAME` named the single state of an `rgbSingle` device `CIE` and
  of an `rgbwSingle` device `RGB`, so neither class bound the state it requires.
