+++ README.md
@@ -160,6 +160,12 @@ Tests are located in the `test/` directory and use ts-node for direct TypeScript
 * (@Apollon77) A Matter fan without an on/off cluster now gets a power state derived from its fan mode, so it can be switched from ioBroker
 * (@Apollon77) Fixed adding one of the newly supported device types from a single state, which named that state wrongly and left the device without its required value
 * (@Apollon77) The controller mapping is now covered by an integration test that commissions a real Matter bridge, and it runs in CI
+* (@Apollon77) The bridge mapping is now covered by a test that replays an exported ioBroker object tree - aliases included - through the type detector and the to-matter converters, and pins the resulting Matter endpoints
+* (@Apollon77) The controller mapping test now also runs the ioBroker type detector over the states a Matter device creates, so every mapped device is pinned to the ioBroker device type those states describe, including the few where they describe none
+* (@Apollon77) Fixed a single-state RGB or RGBW device disappearing from a bridge when its colour state holds a value that is not a hex colour, such as an empty string; an unreadable colour is now reported once per value instead of on every change
+* (@Apollon77) **IMPORTANT:** Fixed the type detection never retrying without the configured type restriction. The retry handed the detector the options object the first attempt had already used, and the detector caches its pattern list and its checked patterns there, so the retry could only ever report nothing. A device whose states do not describe the type it is configured as is now detected as what those states do describe, which is reported as a type mismatch instead of passing silently. Such a device is still exposed as its single configured state, but check the log after updating: a bridge that mixes up a device type now says so
+* (@Apollon77) Fixed a bridged device staying in the bridge when its initialization failed, where it answered controllers with cluster defaults and no ioBroker state behind them; it is now removed from the bridge, as it already was when adding its endpoints failed
+* (@Apollon77) Fixed adding a single-state RGB or RGBW device, which named its state CIE or RGB instead of RGB or RGBW and so left the device without the state it needs
 * (@Apollon77) A thermostat or air conditioner detected only through its heating or cooling setpoint now declares the matching Matter capability, so its setpoint works instead of sitting at the cluster default. Devices already commissioned before this change may need to be re-added in ecosystems that cache the capability set
 * (@Apollon77) Fixed a rejected state write leaving the rejected value cached, and the Device Manager showing a value against the wrong unit
 * (@Apollon77) A bridged thermostat whose ioBroker state declares no temperature range no longer reports a 7 to 30 degree limit to Matter, so setpoints the device itself accepts are no longer rejected
