+++ src/lib/devices/RgbwSingle.ts
@@ -4,6 +4,8 @@ import { type DetectedDevice, type DeviceOptions, StateAccessType } from './Gene
 
 export class RgbwSingle extends Ct {
     #rgbwState?: DeviceStateObject<string>;
+    /** A device stuck on an unreadable colour reports it on every change; log it once per value. */
+    #lastInvalidValue?: string;
 
     constructor(detectedDevice: DetectedDevice, adapter: ioBroker.Adapter, options?: DeviceOptions) {
         super(detectedDevice, adapter, options);
@@ -53,7 +55,13 @@ export class RgbwSingle extends Ct {
         const rgbw = this.#rgbwState.value ?? '#00000000';
         const match = rgbw.match(/^#?([0-9a-f]{2})([0-9a-f]{2})([0-9a-f]{2})([0-9a-f]{2})$/i);
         if (!match) {
-            throw new Error(`Invalid RGBW value: ${rgbw}`);
+            // Anything can write this state, and a value that is not a hex colour is not exceptional;
+            // refusing the whole device over one unreadable colour would drop it out of the bridge.
+            if (this.#lastInvalidValue !== rgbw) {
+                this.#lastInvalidValue = rgbw;
+                this.adapter.log.info(`${this.uuid} Invalid RGBW value: ${rgbw}`);
+            }
+            return { red: 0, green: 0, blue: 0, white: 0 };
         }
         return {
             red: parseInt(match[1], 16),
